🎨  Creates Swift file exclusive to BasePatchRequest
