+++ FlipConnectSDK/PatchRequestProtocol.swift
@@ -13,24 +13,3 @@ protocol PatchRequestProtocol {
     
     func add(operation: Operation, path: EnumType, value: String?)
 }
-
-class BasePatchRequest {
-    var patches: [Patch]? = nil
-    
-    var key: String? = nil
-    
-    func createDictionary() -> JSON {
-        if let patches = patches {
-            var patchesDicts: [JSON] = []
-            
-            for patch in patches {
-                patchesDicts.append(patch.createDictionary())
-            }
-            return [
-                "key": key as Any,
-                "patches": patchesDicts
-            ]
-        }
-        return [:]
-    }
-}
